requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.10.3
bleach==6.1.0
blinker==1.6.2
Brlapi==0.8.4
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.3.1
click==8.1.6
cloud-init==23.3.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
contourpy==1.1.1
cryptography==38.0.4
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
distro==1.8.0
distro-info==1.5
Django==4.2.6
django-phonenumber-field==7.2.0
docutils==0.20.1
fonttools==4.44.0
frozenlist==1.4.0
future==0.18.3
geocoder==1.38.1
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
jaraco.classes==3.2.1
jeepney==0.8.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
keyring==24.2.0
Kivy==2.2.1
Kivy-Garden==0.1.5
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
louis==3.26.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
mdurl==0.1.2
more-itertools==10.1.0
multidict==6.0.4
netifaces==0.11.0
numpy==1.26.1
oauthlib==3.2.2
olefile==0.46
packaging==23.2
pandas==2.1.1
pexpect==4.8.0
phonenumbers==8.13.23
Pillow==10.0.0
ptyprocess==0.7.0
pycairo==1.24.0
pycups==2.0.1
Pygments==2.15.1
PyGObject==3.46.0
PyJWT==2.7.0
pyparsing==3.1.0
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.6.0+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
pytz==2023.3
pyxdg==0.28
PyYAML==6.0.1
ratelim==0.1.6
reportlab==4.0.6
requests==2.31.0
rich==13.3.1
SecretStorage==3.3.3
six==1.16.0
sqlparse==0.4.4
systemd-python==235
twilio==8.10.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.2
unattended-upgrades==0.1
urllib3==1.26.16
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
